+++ secrets/secrets.nix
@@ -1,6 +1,11 @@
 let
   inherit (import ../keys.nix) users hosts;
 in {
+  "acme_zoneee.age".publicKeys = builtins.attrValues {
+    inherit (users) enzime_phi;
+    inherit (hosts) phi;
+  };
+
   "duckdns.age".publicKeys = builtins.attrValues {
     inherit (users) enzime_phi;
     inherit (hosts) phi;
@@ -9,6 +14,11 @@ in {
   "etesync-dav.age".publicKeys = builtins.attrValues {
     inherit (users) enzime_phi;
     inherit (hosts) phi sigma;
+  };
+
+  "nextcloud.age".publicKeys = builtins.attrValues {
+    inherit (users) enzime_phi;
+    inherit (hosts) phi;
   };
 
   "x11vnc_phi.age".publicKeys = builtins.attrValues {